Make find_properties a method on PLYElement, fix column/property terminology
The implementation of `find_properties` is now a method on PLYElement and the method on PLYReader simply calls it on the current element. Both PLYElement and PLYReader have a version of the function which takes a variable-length argument list as per the previous interface. Both call through to `find_properties_va`, which takes a va_list param and does the actual work. The thinking here is that it's useful to be able to test for the existence of properties en-masse before you start parsing. That way you can avoid the cost of scanning through the file if it doesn't contain the data you need.

Renamed `extract_columns` and `extract_list_column` to `extract_properties` and `extract_list_property` respectively. This fixes the confusing mixed terminology that had crept in, where I was sometimes using "property" and sometimes using "column" to refer to the same concept.

Hopefully these will be the last API changes I need to make, at least for a while.

Added NOTES.md with a comment about when to do endianness swapping.