Allow memcpy between signed and unsigned versions of the same base type
Previously we would only memcpy values when the src and dest had exactly the same PLYPropertyType. However we can also use memcpy in the case where one is signed and the other is unsigned, as long as both have the same base type: for example, from int to uint or vice versa. I added in a function to check whether two types are compatible for memcpy purposes and modified the code to use it. This gives a big speed boost on effected files. There aren't that many of them in my test data set, but there's no slowdown on the uneffected files so it's a nice win regardless.