Preallocate memory for list properties
When we load a list property, the data gest stored in a std::vector. Previously this vector was starting out empty, with only the default (small) amount of memory allocated. This meant that we had to repeatedly grow the memory allocation while parsing, which ended up accounting for a significant percentage of the total parsing time.

We now preallocate enough space for every row in the property to be a 3-item list. This is because list properties are usually used for per-face vertex indices and most faces will be triangles, so most of the times lists will be 3 items long. If this doesn't turn out to be enough space, the allocation will still be grown as necessary of course.

This improves performance by about 10% on the pbrt-v3-scenes & benedikt-bitterli collections. The improvement is larger still on the Stanford 3D Scans collection, mainly because of a huge performance increase on one particularly large file (lucy.ply).

lucy.ply represents pretty much the best case for this optimisation: it's a very large model composed entirely of triangular faces. Previously we were having to grow the allocation many times over (24 times, if we assume we start with enough space for 1 item and keep doubling whenever we run out), whereas now we don't have to grow it at all because our initial allocation is the correct size.