Remove hardcoded string length limits
Allocate an additional temporary buffer when constructing a PLYReader.  This (currently) has the same size as the input buffer although it could probably be reduced. We now use this instead of stack-local arrays with hardcoded size for temporary strings. As a result, the only limit on string or identifier length now is that no one token can be longer than the entire input buffer (128 KB).

In `PLYElement::convert_list_to_fixed_size` we still use a stack-local buffer if the generated names will be less than 256 chars long, but in the extremely rare case that they're longer we allocate a temporary buffer on the heap and use that instead. This means we'll avoid having to do any heap allocations 99.999% of the time, but can still handle names that are longer than 256 chars if we encounter any.