Make miniply-info print out the common size for all lists where relevant
`miniply-info` was already printing out a warning when a list property contained lists that aren't all the same size. Now it also prints out a message when they *are* all the same size, saying what that size is. This is handy as a way of finding out whether a model is already triangulated, or all quads, for example.