Add the `PLYReader::sum_of_list_counts` method
Added a new method to return the total number of values in all lists belonging to a given property in the current element. This is useful as a way for user code to calculate the size of the destination array that will be required when calling `extract_list_property`. Previously user code could use the `num_triangles` method where applicable, but otherwise would have to fall back to summing up the row counts.

The result returned by the new method is the same as you would get by summing up the row counts, but is calculated a lot more efficiently. Internally we already know the number of bytes needed to store all of the lists and the number of bytes for a single value in one of the lists, so we can calculate it directly from those instead of iterating.