Add support for loading tri strips as indexed tri sets
Added experimental support for loading triangle strips directly from a PLY file and converting them to an indexed triangle set (i.e. where triangle i is formed from indices 3i, 3i+1 and 3i+2). This is a pair of helper methods in `PLYReader` which interpret the data in a user-specified list property as vertex indices arranged in a triangle strip. The methods are:

`PLYReader::num_triangles_in_strips` which returns the total number of triangles that will be produced from all triangle strips in the property. The main purpose of this is to find out how much space you'll need to allocate in the destination array that you pass in to the next function.

`PLYReader::extract_triangle_strips` produces the list of triangles from the property data.

Both methods support an optional restart value, which is a marker that indicates where one tri strip ends and another begins. The restart marker is optional for the last tri strip in each row: the end of a row is treated as an implicit restart marker if there isn't an actual restart marker present. So there are, effectively, three different ways that the tri strip data can be laid out:

1. One tri strip per row with no explicit restart markers. This uses row boundaries as implicit restart markers.
2. Restart markers used as *separators*, meaning the final tri strip in each row may optionally omit the restart marker.
3. Restart markers used as *terminators*, meaning every tri strip - including the last one in each row - must end with the restart marker.

Layouts 2 and 3 allow for multiple triangle strips to be stored in a single row. The difference is that layout 3 allows us to process the list data more efficiently as a single large block, whereas layout 2 means we have to process it row by row to correctly handle omitted restart markers.