Use correct var name in the triangle mesh sample
PR #3 pointed out some typos in the "how to load a triangle mesh" sample code. One of them was an undefined variable, `propIdx`, when checking whether we need to triangulate the face data. I noticed after merging the PR that it was attempting to work around a typo I'd made in the variable name: it should actually have been `indexes[0]` (or `*indexes`) instead. I'm just fixing that up here.